[WPE][GTK] API test `TestWebProcessExtensions` `/webkit/WebKitWebProcessExtension/user-messages` is flaky https://bugs.webkit.org/show_bug.cgi?id=258883
Reviewed by Philippe Normand.

As explained in `190570@main`, on GTK `webkit_web_view_new` returns
a floating reference, where on WPE it returns an ordinary reference.
There is a special function - `Test::adoptView()` which allows us
not worry about this difference.

Also, this test does a lot of async stuff and can be slow,
so marking it as "slow".

* Tools/TestWebKitAPI/Tests/WebKitGLib/TestWebProcessExtensions.cpp:
(testWebProcessExtensionUserMessages):
* Tools/TestWebKitAPI/glib/TestExpectations.json:

Canonical link: https://commits.webkit.org/265772@main