JSC: Add -fwrapv for GCC and Clang
Left shift of negative value in JSC::RegisterAtOffset::offset()
https://bugs.webkit.org/show_bug.cgi?id=258936

Reviewed by NOBODY (OOPS!).

JSC::RegisterAtOffset::m_offsetBits is ptrdiff_t, so it's signed.  And
on most platforms the stack grows downward, so the value if often
negative.  The C++ standard explicit deems left shift of negative value
undefined.  But in JSC we heavily rely on the wrapping behavior of
integer overflow, so we should use -fwrapv to tell GCC or Clang not to
consider it as undefined.

* Source/JavaScriptCore/CMakeLists.txt:
Add -fwrapv for GCC or Clang.