Fix crash in WebPageProxy::wheelEventHandlingCompleted() https://bugs.webkit.org/show_bug.cgi?id=258705 rdar://111445101
Reviewed by Cameron McCormack.

Crash data show crashes in WebPageProxy::wheelEventHandlingCompleted(), and many of these are under WebProcessProxy::shutDown(),
suggesting that we've closed the WebPageProxy. This is indeed possible, since WebPageProxy::sendWheelEvent() sends IPC
with an async reply.

So early return in the reply handler if closed, and also null-check the m_pageClient (which is a WeakPtr).

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::sendWheelEvent):
(WebKit::WebPageProxy::wheelEventHandlingCompleted):

Canonical link: https://commits.webkit.org/265817@main