[UI-side compositing] Optimize setting NSScrollerImp members in ScrollerMac::updateValues https://bugs.webkit.org/show_bug.cgi?id=258904 rdar://111816135
Reviewed by NOBODY (OOPS!).

Limit the amount of interaction we do with NSScrollerImp in ScrollerMac::updateValues by
keeping track of the last value we set on the NSScrollerImp and not setting the value
if it was the same as the previous value. Accomplish this by maintaining a struct that
holds the last values set on the NSScrollerImp.

* Source/WebCore/page/scrolling/mac/ScrollerMac.h:
* Source/WebCore/page/scrolling/mac/ScrollerMac.mm:
(WebCore::ScrollerMac::updateValues):