REGRESSION(264617@main): fast/forms/ios/file-upload-panel-accept.html is a constant timeout https://bugs.webkit.org/show_bug.cgi?id=258901 rdar://110879270
Reviewed by Simon Fraser.

* Source/WebKit/UIProcess/RemoteLayerTree/RemoteLayerTreeDrawingAreaProxy.mm:
(WebKit::RemoteLayerTreeDrawingAreaProxy::commitLayerTree):
The explicit commit introduced in 264617@main caused multiple bits of unexpected
fallout, breaking a layout test and bypassing commit-coalescing optimizations
in UIKit.

Instead of explicitly committing, just hold on to the IOSurface ports until
the commit arrives in the render server. (We can't hold on to the whole
transaction because the ports are consumed when applying the backing store
to its layer).

Canonical link: https://commits.webkit.org/265825@main