Use composited masking for clip-path which uses an SVG path reference https://bugs.webkit.org/show_bug.cgi?id=258947 rdar://111870932
Reviewed by NOBODY (OOPS!).

stackoverflow.com has a clip-path on the hero element which references an SVG path. That hero element
is also painted on every frame because of some SVG animations, and is composited.

Currently for this configuration we make a masking GraphicsLayer for the clip-path, and paint the
RenderSVGResourceClipper into it to generate the mask. But we can do this more efficiently if
we can construct a CAShapeLayer for the mask, which can can in the same situtations that RenderSVGResourceClipper
can do path-only clipping.

So refactor RenderSVGResourceClipper code a little to make it easiser to fetch the single SVGGraphicsElement
that provides the path, and add a pathOnlyClip() function that returns a pair of the path and a winding rule.

Use that in RenderLayerBacking::updateMaskingLayer() and RenderLayerBacking::updateMaskingLayerGeometry()
to configure the shape layer-based masking.

* LayoutTests/compositing/masks/clip-path-composited-descendent-expected.txt:
* LayoutTests/compositing/masks/compositing-clip-path-change-no-repaint-expected.txt:
* LayoutTests/platform/mac/svg/W3C-SVG-1.1/metadata-example-01-b-expected.txt:
* LayoutTests/platform/mac/svg/custom/use-on-symbol-inside-pattern-expected.txt:
* LayoutTests/svg/clip-path/clip-path-objectboundingbox-003.svg:
* Source/WebCore/platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::GraphicsLayerCA::dumpAdditionalProperties const):
* Source/WebCore/rendering/PathOperation.h:
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::canUsePathClip const):
(WebCore::RenderLayer::computeClipPath const):
(WebCore::RenderLayer::setupClipPath):
* Source/WebCore/rendering/RenderLayer.h:
* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::updateMaskingLayerGeometry):
(WebCore::RenderLayerBacking::updateMaskingLayer):
* Source/WebCore/rendering/svg/RenderSVGResourceClipper.cpp:
(WebCore::RenderSVGResourceClipper::pathOnlyClipSourceElement const):
(WebCore::RenderSVGResourceClipper::canUsePathOnlyClip const):
(WebCore::RenderSVGResourceClipper::pathOnlyClip const):
(WebCore::RenderSVGResourceClipper::applyPathOnlyClipping const):
(WebCore::RenderSVGResourceClipper::applyClippingToContext):
(WebCore::RenderSVGResourceClipper::pathOnlyClipping): Deleted.
* Source/WebCore/rendering/svg/RenderSVGResourceClipper.h: