[WPE] The bots WPE-Linux-RPi4-32bits-Mesa-Release-Perf-Tests are failing to autoinstall the python dependency bcrypt https://bugs.webkit.org/show_bug.cgi?id=258941
Reviewed by Philippe Normand.

There are no pre-built packages on the python repository for ARM-32 arcitecture
for the depency bcrypt. And installing it from source is a major headache because
this dependency requires a cargo/rust toolchain that is complicated to add into
the yocto-based image that the bots execute.

The dependency itself (bcrypt) is not needed at all in the bots so if we avoid
trying to autoinstall it things work as expected.

But for that we have also to avoid autoinstalling as well the cryptography python
dependency, which also depends on a cargo/rust toolchain and has no prebuilt
version for ARM-32 bits.

I tested to run the usual python tools: run-webkit-tests, run-minibrowser, run-benchmark
and everything works as expected without installing this two dependencies.

The architecture name is not unique and can be at least any of this: 'armv7l', 'armv7', 'armv8l'
so to match it we check if starts with 'arm' and is '32bit'

* Tools/Scripts/libraries/webkitcorepy/webkitcorepy/__init__.py:
* Tools/Scripts/webkitpy/autoinstalled/twisted.py:

Canonical link: https://commits.webkit.org/265835@main