AX VoiceOver: some content on web pages is not displayed on the braille display. https://bugs.webkit.org/show_bug.cgi?id=259066 rdar://110758833
Reviewed by Tyler Wilcock.

AX clients like VoiceOver may request the AttributedString of a TextMarkerRange using the WebArea object. Since we don't cache the AttributedString for the WebArea, this causes a hit to the main thread. Furthermore, since the TextMarkerRange for a given object is created off the main thread and expects to be accessed off the main thread, it may not return the correct AttributedString when used on the main thread. This change ensures that the cached AttributedString for a given object is returned even when it is requested using the WebArea or any other object instead the object that owns the range.

Canonical link: https://commits.webkit.org/265957@main