SQLiteDatabase should try to set RunningBoard file attribute on all database files, not just the main one https://bugs.webkit.org/show_bug.cgi?id=259117 rdar://97212284
Reviewed by Ben Nham.

SQLiteDatabase should try to set RunningBoard file attribute on all database
files, not just the main one.

This is a follow-up to 265951@main. This is not perfect this we don't control
when the `-shm` file will be created for example. However, this is a best
effort until we have something better (like an API from SQLite).

* Source/WebCore/platform/sql/SQLiteDatabase.cpp:
(WebCore::SQLiteDatabase::open):
* Source/WebCore/platform/sql/SQLiteFileSystem.cpp:
(WebCore::SQLiteFileSystem::setCanSuspendLockedFileAttribute):
* Source/WebCore/platform/sql/SQLiteFileSystem.h:

Canonical link: https://commits.webkit.org/265962@main